requirements.txt
Django==4.2.4
django-import-export==3.3.6
Pillow==10.0.0
requests==2.31.0
